Code updated, sample data file moved to app
